two divs, one holds chart, the other holds the methane map
data1.csv & data2.csv are samples of us state abbreviations with some random num data

next step is to use bootstrap and adjust the columns to place the methane map to the left and the chart to the right.

After that need to get the state names to appear in the Y axis.

This is a stopping point for now until. I have to go give a final presentation for a different class.