got div to hold repestive methane map and skeleton of chart.
There's a couple approaches to organizing the format of the view. 1) bootstrap classes on each div, 2) hardcore svg width & height so that it fills the div appropriately.

I'd like to make it responsive so there must be some way to get the svg's width & height to adapt to the width and height of the div.

I tried putting a class on each div and that seems to help in sizing / positiing the divs, however, svg does not respond to those changes so there must be a better way