geting data to appear on the graph
I'm having trouble getting the data to appear on the graph.  The structure of the graph is working.

I'm accessing the data with this func:

    // Get the data
    d3.csv("data2.csv", function (error, data) {
      data.forEach(function (d) {

        console.log(d.state);
        console.log(d.value);

        d.value = +d.value

        return d.value;
      });

console logs give me what I want.

this line d.value = +d.value;

is a best practice for ensuring that the data returned in numerical - which it is.

taking a break not sure how to proceed.

been reading these docs:

https://leanpub.com/D3-Tips-and-Tricks/read#leanpub-auto-setting-up-the-margins-and-the-graph-area